Changed continuous format from doubles to floats.
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@534 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/format/DataFormatter.java
@@ -197,7 +197,7 @@ public class DataFormatter implements FormatProgressListener {
         BufferedReader inFile = this.openInputFile();
 
         List<FieldType> fields = new ArrayList<FieldType>();
-        fields.add(FieldType.DOUBLE);
+        fields.add(FieldType.FLOAT);
 
         List<Object> modifiers = new ArrayList<Object>();
         modifiers.add(null);
